Restore green CI rehearsals (#70)
Cast the environment-backed password timeout and install the locked browser runtime so application checks pass on a clean runner. Supply root-owned rehearsal credentials and a host-publishable PostgreSQL network so healthy, restart, and rollback rehearsals complete with masked diagnostics.

+++ config/auth.php
@@ -112,6 +112,6 @@ return [
     |
     */
 
-    'password_timeout' => env('AUTH_PASSWORD_TIMEOUT', 900),
+    'password_timeout' => (int) env('AUTH_PASSWORD_TIMEOUT', 900),
 
 ];
